mega-fix: Fazer com que o parse e criacao de elementos seja feita ao input, ou seja, a cada nova letra digitada sem erros ou crashs

+++ scripts/interpreter/interpreter.js
@@ -49,7 +49,7 @@ class Interpreter {
       const element = entry[1];
       this.elements[entry[0]] = null;
 
-      element.delete();
+      if (element) element.delete();
     });
   }
 }
